requirements.txt
beautifulsoup4
skyfield
PyOpenGL
glfw
pyglm
Pillow
